requirements.txt
Click
litellm
python-dotenv
PyPDF2
python-docx
python-pptx
pandas
pyarrow
loguru
tenacity
tqdm
openpyxl
datasets
